Rename node extension utils for clarity
finding the file with sync IO utils is tricky because it is called `utils`. this change gives the file a more clear name.
