+++ packages/next/src/server/node-environment-extensions/random.tsx
@@ -6,7 +6,7 @@
  * The extensions here never error nor alter the random generation itself and thus should be transparent to callers.
  */
 
-import { io } from './utils'
+import { io } from './io-utils'
 
 const expression = '`Math.random()`'
 try {
